Removed extras module and moved to core
