+++ meteo-core/src/main/java/no/api/meteo/service/locationforecast/extras/LocationForecastHelper.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package no.api.meteo.services;
+package no.api.meteo.service.locationforecast.extras;
 
 import lombok.NonNull;
 import no.api.meteo.MeteoException;
