+++ meteo-core/src/main/java/no/api/meteo/service/locationforecast/extras/ScoreForecast.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package no.api.meteo.services;
+package no.api.meteo.service.locationforecast.extras;
 
 import lombok.Value;
 import no.api.meteo.entity.core.service.locationforecast.PeriodForecast;
