+++ meteo-core/src/test/java/no/api/meteo/service/locationforecast/extras/MeteoForecastIndexerTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package no.api.meteo.services;
+package no.api.meteo.service.locationforecast.extras;
 
 import no.api.meteo.MeteoException;
 import no.api.meteo.entity.core.service.locationforecast.LocationForecast;
