--- meteo-extras/pom.xml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>meteo</artifactId>
-        <groupId>no.api.meteo</groupId>
-        <version>3.0.0-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>meteo-extras</artifactId>
-    <name>meteo-extras</name>
-    <description>Meteo Extras package</description>
-
-    <profiles>
-        <profile>
-            <id>jrebel</id>
-            <activation>
-                <property>
-                    <name>env</name>
-                    <value>jrebel</value>
-                </property>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.zeroturnaround</groupId>
-                        <artifactId>jrebel-maven-plugin</artifactId>
-                        <executions>
-                            <execution>
-                                <id>generate-rebel-xml</id>
-                                <phase>process-resources</phase>
-                                <goals>
-                                    <goal>generate</goal>
-                                </goals>
-                            </execution>
-                        </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-    </profiles>
-
-
-    <dependencies>
-        <dependency>
-            <groupId>no.api.meteo</groupId>
-            <artifactId>meteo-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>no.api.meteo</groupId>
-            <artifactId>meteo-test</artifactId>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-
-</project>
